@@ -17,7 +17,7 @@
     "@reduxjs/toolkit": "1.9.2",
     "@testing-library/dom": "7.21.4",
     "dotenv": "^16.0.3",
-    "next": "13.1.6",
+    "next": "13.4.12",
     "next-mdx-remote": "^4.3.0",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
