@@ -1,12 +1,11 @@
 import { NavBarIcon } from "./NavBarIcon";
+import Link from "next/link";
 import { SiHomeadvisor } from "react-icons/si";
 import { MdArticle, MdCatchingPokemon } from "react-icons/md";
 import { AiTwotoneHome } from "react-icons/ai";
 import { BsNewspaper, BsInfoCircleFill } from "react-icons/bs";
-import { useRouter } from "next/router";
 
 export default function NavBar() {
-  const router = useRouter();
   const tabBarTitle = "text-sm";
   const tabBarButton =
     "flex flex-col rounded border border-gray-900 p-1 items-center text-purple-100 bg-gray-800";
@@ -20,24 +19,28 @@ export default function NavBar() {
               "grid grid-flow-col justify-stretch space-x-1 bg-gray-600 rounded border-2 p-1 border-purple-100"
             }
           >
-            <button className={tabBarButton} onClick={() => router.push("/")}>
+            <Link href={'/'} passHref>
+            <button className={tabBarButton}>
               <AiTwotoneHome size="25" />
               <div className={tabBarTitle}>Home</div>
             </button>
+            </Link>
+            <Link href={"/pokemon"} passHref>
             <button
               className={tabBarButton}
-              onClick={() => router.push("/pokemon")}
             >
               <MdCatchingPokemon size="25" />
               <div className={tabBarTitle}>Pokemon</div>
             </button>
+            </Link>
+            <Link href={"/articles"} passHref>
             <button
               className={tabBarButton}
-              onClick={() => router.push("/articles")}
             >
               <BsNewspaper size="25" />
               <div className={tabBarTitle}>Articles</div>
             </button>
+            </Link>
           </div>
         </div>
       </div>
